[VGNWB-351] Refactor Repository URLs
