+++ annotations/views/__init__.py
@@ -1,4 +1,4 @@
 from . import \
     project_views, \
     quadruple_views, relationtemplate_views, rest_views, \
-    user_views, repository_views, annotation_views
+    user_views, repository_views, annotation_views, amphora_views, citesphere_views
